clients/python/requirements.txt
paho-mqtt==1.6.1
jsonschema==4.23.0
pytest==8.3.2
